dns: rename rust files and funcs
